Client: Added File Download and Offer Services
- Introduced new services for file download and offer management.
- The file download service handles data chunks, cancels downloads, and manages upload/download cancellations.
- It also provides progress tracking for ongoing downloads.
- Both services extend from a base file transfer service, ensuring consistent behavior across different types of transfers.
